==> 2group/preprocessing/STEP2_merge_meta.ipynb <==
## Merge meta retracted

Connect to drive, set directory

from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/Thesis_UU/push/2group/Data (CSV)'

Load files and merge

import pandas as pd

df_11_journals = pd.read_csv('retracted_texts_11_journals.csv')
df_meta_11_journals = pd.read_csv('meta_retracted_11_journals.csv')

df_11_journals['ID'] = df_11_journals.DOI.str.replace('.pdf', '', regex = True) #remove ".pdf" from ID

r_result = pd.merge(df_11_journals, df_meta_11_journals, on="ID") #merge meta

r_result.Journal.value_counts() #number of papers per journal

## Merge meta non-Retracted

Load files and merge

import pandas as pd

df_11_journals_non_retracted = pd.read_csv('non_retracted_11_journals.csv')
df_meta_11_journals_non_retracted = pd.read_csv('meta_non_retracted_11_journals.csv')

df_11_journals_non_retracted['ID'] = df_11_journals_non_retracted.DOI.str.replace('.pdf', '', regex = True)

nr_result = pd.merge(df_11_journals_non_retracted, df_meta_11_journals_non_retracted, on="ID")

nr_result['Source Title'].value_counts() #number of papers per journal

Save files to csv

nr_result.to_csv('non_retracted_11_journals_w_meta.csv')
r_result.to_csv('retracted_11_journals_w_meta.csv')
